Restructured files to hopefully fix foreseeable installation issues; added a simple GUI with minimal additions to handle button events. Errors still exist within DatabaseAccess.cs
