--- CEIS400-ECS/CEIS400-ECS.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.36401.2 d17.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CEIS400-ECS", "CEIS400-ECS.csproj", "{7FFAEECD-5D8E-4D74-88AE-3FF9A56EDCE3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7FFAEECD-5D8E-4D74-88AE-3FF9A56EDCE3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7FFAEECD-5D8E-4D74-88AE-3FF9A56EDCE3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7FFAEECD-5D8E-4D74-88AE-3FF9A56EDCE3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7FFAEECD-5D8E-4D74-88AE-3FF9A56EDCE3}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {DF504ED5-8092-4398-A530-45A5D83391B9}
-	EndGlobalSection
-EndGlobal
