--- CEIS400-Equipment-Checkout-System.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.36310.24 d17.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CEIS400-Equipment-Checkout-System", "CEIS400-Equipment-Checkout-System.csproj", "{0CDF8C47-4C03-44D1-ABEE-C624307FDEB5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0CDF8C47-4C03-44D1-ABEE-C624307FDEB5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0CDF8C47-4C03-44D1-ABEE-C624307FDEB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0CDF8C47-4C03-44D1-ABEE-C624307FDEB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0CDF8C47-4C03-44D1-ABEE-C624307FDEB5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8703CA94-3B72-459C-875C-3370BCC553F6}
-	EndGlobalSection
-EndGlobal
