+++ CEIS400_ECS.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.14.36408.4
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CEIS400_ECS", "CEIS400_ECS.csproj", "{7F7BA2F2-C7E0-455C-B65A-BB6940E93CE8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7F7BA2F2-C7E0-455C-B65A-BB6940E93CE8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7F7BA2F2-C7E0-455C-B65A-BB6940E93CE8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7F7BA2F2-C7E0-455C-B65A-BB6940E93CE8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7F7BA2F2-C7E0-455C-B65A-BB6940E93CE8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {FB4FCA67-6851-43C3-9001-5A64A05ED449}
+	EndGlobalSection
+EndGlobal
