+++ IHasBarcode.cs
